Rename folder to add kratos in the name
